:bug: Fixed bugs and perm issues

+++ src/commands/admin/unsticky.js
@@ -1,20 +1,27 @@
 const { setStickyData } = require("../../utils/functions");
 
 module.exports = {
-    name: "unsticky",
-    description: "Sticky a message to the bottom of the screen",
-    aliases: ["removesticky"],
-    category: "admin",
-    execute(bot, message) {
-        message.delete();
+  name: "unsticky",
+  description: "Sticky a message to the bottom of the screen",
+  aliases: ["removesticky"],
+  category: "admin",
+  execute(bot, message) {
+    if (!message.guild.me.hasPermission("MANAGE_MESSAGES"))
+      return message.reply(
+        "I don't have the correct permissions to manage messages! (Manage Messages)"
+      );
 
-        const member = message.member;
+    message.delete();
 
-        if (!member.hasPermission("MANAGE_MESSAGES"))
-            return message.reply("You don't have the correct permission! (Manage messages)");
+    const member = message.member;
 
-        setStickyData(message.guild.id, {});
+    if (!member.hasPermission("MANAGE_MESSAGES"))
+      return message.reply(
+        "You don't have the correct permission! (Manage messages)"
+      );
 
-        message.channel.send(`Cleared sticky for ${message.channel.name}`);
-    }
+    setStickyData(message.guild.id, {});
+
+    message.channel.send(`Cleared sticky for ${message.channel.name}`);
+  },
 };
